cli/console: Split ssh client code
